Fixed a bug that didn't let me run the program on Windows 10
I found a similar error in Stack Overflow (https://stackoverflow.com/q/18204782)
and implemented the best solution. The problem seems to be related
to recursive subprocesses in Windows. I think the solution shouln't
affect other OS, but it can be a good idea to run the new script on
other OS before accepting a pull request.

Signed-off-by: jbeleno <jsbeleno@gmail.com>